+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Pretty-printing of expressions in REPL (#870)
 - Spread syntax for record updates (#880)
+- Save ITF traces in `quint test` (#884)
 
 ### Changed
 ### Deprecated